refactor: Profits 일부 기능 Profit에 위임
