+++ src/main/java/blackjack/domain/machine/Profits.java
@@ -8,8 +8,6 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 
 public class Profits {
-    private static final double BLACKJACK_RATE = 1.5;
-
     private final Map<Participant, Profit> result;
 
     private Profits(Map<Participant, Profit> result) {
@@ -19,49 +17,11 @@ public class Profits {
     public static Profits of(Dealer dealer, Players players) {
         Map<Participant, Profit> result = new LinkedHashMap<>();
         for (Player player : players.getPlayers()) {
-            result.put(player, getProfit(dealer, player));
+            result.put(player, Profit.of(dealer, player));
         }
 
-        result.put(dealer, getDealerProfit(result));
+        result.put(dealer, Profit.getDealerProfit(result));
         return new Profits(result);
-    }
-
-    private static Profit getProfit(Dealer dealer, Player player) {
-        if (player.initScore().isMax()) {
-            return getProfitForBlackjack(dealer, player);
-        }
-
-        return getOrdinaryProfit(dealer, player);
-    }
-
-    private static Profit getOrdinaryProfit(Dealer dealer, Player player) {
-        Record record = Record.getRecord(player, dealer);
-
-        if (record == Record.VICTORY) {
-            return new Profit(player.getBetting().getMoney());
-        }
-
-        if (record == Record.DEFEAT) {
-            return new Profit(-player.getBetting().getMoney());
-        }
-
-        return new Profit();
-    }
-
-    private static Profit getProfitForBlackjack(Dealer dealer, Player player) {
-        if (dealer.initScore().isMax()) {
-            return new Profit(player.getBetting().getMoney());
-        }
-
-        return new Profit((long) (BLACKJACK_RATE * player.getBetting().getMoney()));
-    }
-
-    private static Profit getDealerProfit(Map<Participant, Profit> playersProfit) {
-        long sum = playersProfit.values().stream()
-                .mapToLong(Profit::getMoney)
-                .sum();
-
-        return new Profit(-sum);
     }
 
     public Map<Participant, Profit> getResult() {
